requirements.txt
streamlit
sentence-transformers
faiss-cpu
pypdf
numpy
torch